# contracts/mvp/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Планы? MVP API
  version: "0.1.0"
  description: Contract-first MVP backend for the Планы? / Fest&Rest app.
servers:
  - url: http://localhost:3001/api
    description: Local development
security: []
...
paths:
...
  /users/me:
    get:
      tags: [users]
      operationId: getUserMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    patch:
      tags: [users]
      operationId: patchUserMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/{id}:
    get:
      tags: [users]
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /users/friends:
    get:
      tags: [users]
      operationId: listFriends
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/FriendshipStatus"
      responses:
        "200":
          description: Friend list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendListResponse"
  /users/friends/{id}:
    post:
      tags: [users]
      operationId: addFriend
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Friendship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendshipResponse"
    patch:
      tags: [users]
      operationId: respondToFriendRequest
      description: >-
        Accept or decline a pending friend request. Caller must be the
        addressee. On `accept`, a `friend_accepted` notification is created
        for the original requester with payload
        `{friendship_id, accepter_id, accepter_name, accepter_username}`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FriendActionRequest"
      responses:
        "200":
          description: Friendship accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendshipResponse"
        "204":
          description: Friendship declined
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [users]
      operationId: removeFriend
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "204":
          description: Friendship removed
  /users/search:
    get:
      tags: [users]
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Case-insensitive substring match against name/username.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Matching users, with current-user friendship status embedded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string

    # --- Enums ---
...
    FriendshipStatus:
      type: string
      enum: [pending, accepted]
...
    User:
      type: object
      required: [id, phone, name, username, created_at]
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
        username:
          type: string
        avatar_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    PatchUserRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        avatar_url:
          type: string
...
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    FriendListResponse:
      type: object
      required: [friends]
      properties:
        friends:
          type: array
          items:
            $ref: "#/components/schemas/User"
    FriendshipResponse:
      type: object
      required: [friendship]
      properties:
        friendship:
          type: object
          required: [id, requester_id, addressee_id, status, created_at]
          properties:
            id:
              type: string
              format: uuid
            requester_id:
              type: string
              format: uuid
            addressee_id:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/FriendshipStatus"
            created_at:
              type: string
              format: date-time
...
    FriendActionRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, decline]
    UserWithFriendship:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            friendship_status:
              type: string
              nullable: true
              enum: [friend, request_sent, request_received]
              description: >-
                Current-viewer friendship state with this user, if any.
                `friend` = accepted; `request_sent` = viewer sent a pending
                request; `request_received` = the other user sent a pending
                request to the viewer. `null` means no relationship.
    UserSearchResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserWithFriendship"
...